sendmax work + changes to carry a tx
Updated the getExchangeRates and initSwap methods to account for a tx parameter instead of a raw amount. This enables a possibly imperfect send max (in the sense that it might leave some dust) that will be used on desktop/mobile.

Also updated the cli command to be able to sendMax for testing.